debugging

+++ src/vector.rs
@@ -10,7 +10,7 @@ pub struct Vector {
     pub values: [Float; 1326],
 }
 
-pub(crate) type Float = f32;
+pub(crate) type Float = f64;
 
 impl Vector {
     pub fn from(v: &[Float]) -> Self {
